+++ CHANGELOG.md
@@ -15,6 +15,7 @@
   * Now parses `Position.Assessment` (see above).
 * Add `Move.Assessment(notation:)` initializer.
 * Add `all` property to `Game.Tags` to return array of all named tags.
+* Add `CustomStringConvertible` conformance to `Game` and `Move`.
 
 ### Bug Fixes
 * Fix `SANParser` not parsing en passant captures properly (by Rob Raese).